Added ModLib.Definitions substitute
Refactoring IPropertyDefinition and ISettingsPropertyDefinition, still in progress
Added the old MCM compatibility .dll's for now
Added the e1.3.1 and e1.4.0 as supported, will be refactored
Other minor stuff like namespace moving

+++ MCM.UI/GUI/ViewModels/EditValueVM.cs
@@ -51,9 +51,9 @@ namespace MCM.UI.GUI.ViewModels
         [DataSourceProperty]
         public SettingType SettingType => SettingProperty.SettingType;
         [DataSourceProperty]
-        public float MinValue => SettingProperty.SettingPropertyDefinition.EditableMinValue;
+        public float MinValue => (float) SettingProperty.SettingPropertyDefinition.EditableMinValue;
         [DataSourceProperty]
-        public float MaxValue => SettingProperty.SettingPropertyDefinition.EditableMaxValue;
+        public float MaxValue => (float) SettingProperty.SettingPropertyDefinition.EditableMaxValue;
 
         public EditValueVM(SettingsPropertyVM settingProperty)
         {
